chore(system): extract queue consumers processing into message handlers
