+++ executor/src/main/java/io/kestra/executor/ExecutionStateStore.java
@@ -2,6 +2,7 @@ package io.kestra.executor;
 
 import io.kestra.core.models.executions.Execution;
 
+import java.util.Optional;
 import java.util.function.Function;
 
 /**
@@ -11,5 +12,5 @@ public interface ExecutionStateStore {
     /**
      * Lock an execution for processing using the provided function.
      */
-    ExecutorContext lock(String executionId, Function<Execution, ExecutorContext> function);
+    Optional<ExecutorContext> lock(String executionId, Function<Execution, ExecutorContext> function);
 }
