+++ executor/src/test/java/io/kestra/executor/handler/WorkerTaskResultMessageHandlerTest.java
@@ -1,0 +1,87 @@
+package io.kestra.executor.handler;
+
+import io.kestra.core.junit.annotations.KestraTest;
+import io.kestra.core.models.executions.Execution;
+import io.kestra.core.models.executions.TaskRun;
+import io.kestra.core.models.flows.GenericFlow;
+import io.kestra.core.models.flows.State;
+import io.kestra.core.repositories.ExecutionRepositoryInterface;
+import io.kestra.core.repositories.FlowRepositoryInterface;
+import io.kestra.core.runners.WorkerTaskResult;
+import jakarta.inject.Inject;
+import org.junit.jupiter.api.Test;
+
+import java.util.Collections;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@KestraTest
+class WorkerTaskResultMessageHandlerTest {
+    @Inject
+    private WorkerTaskResultMessageHandler workerTaskResultMessageHandler;
+
+    @Inject
+    private ExecutionRepositoryInterface executionRepository;
+
+    @Inject
+    private FlowRepositoryInterface flowRepository;
+
+    @Test
+    void shouldReturnEmptyForNonExistingExecution() {
+        var workerTaskResult = WorkerTaskResult.builder()
+            .taskRun(TaskRun.builder()
+                .executionId("execution")
+                .id("taskrun")
+                .taskId("task")
+                .build())
+            .build();
+
+        var maybeExecutor = workerTaskResultMessageHandler.handle(workerTaskResult);
+
+        assertThat(maybeExecutor).isEmpty();
+    }
+
+    @Test
+    void shouldReturnAnExecutorForExistingExecution() {
+        var flow = Fixtures.flow();
+        flowRepository.create(GenericFlow.of(flow));
+        var execution = Execution.newExecution(flow, Collections.emptyList());
+        var taskRun = TaskRun.builder()
+            .executionId(execution.getId())
+            .namespace(execution.getNamespace())
+            .flowId(execution.getFlowId())
+            .id("taskrun")
+            .taskId(flow.getTasks().getFirst().getId())
+            .state(new State().withState(State.Type.SUBMITTED))
+            .build();
+        executionRepository.save(execution.withTaskRunList(Collections.singletonList(taskRun)));
+        var workerTaskResult = WorkerTaskResult.builder()
+            .taskRun(taskRun.withState(State.Type.SUCCESS))
+            .build();
+
+        var maybeExecutor = workerTaskResultMessageHandler.handle(workerTaskResult);
+
+        assertThat(maybeExecutor).isPresent();
+        assertThat(maybeExecutor.get().getExecution().getTaskRunList().getFirst().getState().getCurrent()).isEqualTo(State.Type.SUCCESS);
+    }
+
+    @Test
+    void shouldFailTheExecutionForMissingTask() {
+        var flow = Fixtures.flow();
+        flowRepository.create(GenericFlow.of(flow));
+        var execution = Execution.newExecution(flow, Collections.emptyList());
+        executionRepository.save(execution);
+        var workerTaskResult = WorkerTaskResult.builder()
+            .taskRun(TaskRun.builder()
+                .executionId(execution.getId())
+                .id("taskrun")
+                .taskId("task")
+                .build())
+            .build();
+
+        var maybeExecutor = workerTaskResultMessageHandler.handle(workerTaskResult);
+
+        assertThat(maybeExecutor).isPresent();
+        assertThat(maybeExecutor.get().getExecution().getState().getCurrent()).isEqualTo(State.Type.FAILED);
+    }
+}
